@@ -3,10 +3,11 @@
   "version": "2.6.1",
   "description": "Service for DSCP",
   "main": "app/index.js",
+  "type": "module",
   "scripts": {
-    "test": "NODE_ENV=test mocha --config ./test/mocharc.js ./test",
-    "test:integration": "NODE_ENV=test mocha --config ./test/mocharc.js ./test/integration",
-    "test:unit": "NODE_ENV=test mocha --config ./test/mocharc-unit.js ./test/unit",
+    "test": "NODE_ENV=test mocha --config ./test/mocharc.cjs ./test",
+    "test:integration": "NODE_ENV=test mocha --config ./test/mocharc.cjs ./test/integration",
+    "test:unit": "NODE_ENV=test mocha --config ./test/mocharc-unit.cjs ./test/unit",
     "lint": "eslint .",
     "depcheck": "depcheck",
     "start": "node app/index.js",
@@ -31,29 +32,29 @@
   },
   "homepage": "https://github.com/digicatapult/dscp-ipfs#readme",
   "dependencies": {
-    "@digicatapult/dscp-node": "^4.2.1",
-    "dotenv": "^8.2.0",
-    "envalid": "^7.0.0",
-    "express": "^4.17.1",
-    "pino": "^6.11.1",
-    "pino-http": "^5.5.0"
+    "@digicatapult/dscp-node": "^4.3.1",
+    "dotenv": "^16.0.3",
+    "envalid": "^7.3.1",
+    "express": "^4.18.2",
+    "pino": "^8.6.1",
+    "pino-http": "^8.2.1"
   },
   "devDependencies": {
-    "@babel/eslint-parser": "^7.18.9",
-    "chai": "^4.3.1",
+    "@babel/eslint-parser": "^7.19.1",
+    "chai": "^4.3.6",
     "delay": "^5.0.0",
-    "depcheck": "^1.4.0",
-    "eslint": "^8.13.0",
-    "eslint-config-prettier": "^8.1.0",
-    "eslint-plugin-prettier": "^3.3.1",
-    "form-data": "^3.0.1",
-    "go-ipfs": "^0.13.1",
-    "mocha": "^9.2.0",
-    "node-fetch": "npm:@achingbrain/node-fetch@^2.6.7",
-    "nodemon": "^2.0.19",
+    "depcheck": "^1.4.3",
+    "eslint": "^8.25.0",
+    "eslint-config-prettier": "^8.5.0",
+    "eslint-plugin-prettier": "^4.2.1",
+    "formdata-node": "^5.0.0",
+    "go-ipfs": "^0.16.0",
+    "mocha": "^10.1.0",
+    "node-fetch": "^3.2.10",
+    "nodemon": "^2.0.20",
     "nyc": "^15.1.0",
-    "pino-colada": "^2.1.0",
-    "prettier": "^2.2.1",
-    "sinon": "^12.0.1"
+    "pino-colada": "^2.2.2",
+    "prettier": "^2.7.1",
+    "sinon": "^14.0.1"
   }
 }
