@@ -1,10 +1,18 @@
-const path = require('path')
-const envalid = require('envalid')
-const dotenv = require('dotenv')
+import path from 'path'
+import envalid from 'envalid'
+import dotenv from 'dotenv'
+import { fileURLToPath } from 'url'
+import os from 'os'
+import fsWithSync from 'fs'
+
+const __filename = fileURLToPath(import.meta.url)
+const __dirname = path.dirname(__filename)
 
 /* istanbul ignore else */
 if (process.env.NODE_ENV === 'test') {
   dotenv.config({ path: 'test/test.env' })
+  const ipfsDir = fsWithSync.mkdtempSync(path.join(os.tmpdir(), 'dscp-ipfs-'))
+  process.env.IPFS_PATH = ipfsDir
 }
 
 const validateArgs = envalid.makeValidator((input) => {
@@ -28,7 +36,7 @@ const vars = envalid.cleanEnv(
       default: 'ipfs',
       devDefault: path.resolve(__dirname, '..', `node_modules`, '.bin', 'ipfs'),
     }),
-    IPFS_ARGS: validateArgs({ default: '["daemon", "--migrate"]' }),
+    IPFS_ARGS: validateArgs({ default: ['daemon', '--migrate'] }),
     IPFS_LOG_LEVEL: envalid.str({ default: 'info', devDefault: 'debug' }),
     HEALTHCHECK_POLL_PERIOD_MS: envalid.num({ default: 30 * 1000, devDefault: 1000 }),
     HEALTHCHECK_TIMEOUT_MS: envalid.num({ default: 2 * 1000, devDefault: 1000 }),
@@ -37,5 +45,4 @@ const vars = envalid.cleanEnv(
     strict: true,
   }
 )
-
-module.exports = vars
+export default vars
