@@ -1,31 +1,26 @@
-const { createHttpServer } = require('../../../app/server')
-const logger = require('../../../app/logger')
-const { PORT } = require('../../../app/env')
+import { createHttpServer } from '../../../app/server.js'
+import logger from '../../../app/logger.js'
+import env from '../../../app/env.js'
 
-async function startServer(context) {
+export async function startServer(context) {
   const create = await createHttpServer()
   await new Promise((resolve, reject) => {
     context.ipfs = create.ipfs
-    context.server = create.app.listen(PORT, (err) => {
+    context.server = create.app.listen(env.PORT, (err) => {
       if (err) {
         logger.error('Error starting app:', err)
         reject(err)
       } else {
-        logger.info(`Server is listening on port ${PORT}`)
+        logger.info(`Server is listening on port ${env.PORT}`)
         resolve()
       }
     })
   })
 }
 
-async function stopServer(context) {
+export async function stopServer(context) {
   await context.ipfs.stop()
   await new Promise((resolve) => {
     context.server.close(() => resolve())
   })
-}
-
-module.exports = {
-  startServer,
-  stopServer,
 }
